geoobra_backend_v3/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
SQLAlchemy>=2.0
alembic>=1.13
pydantic>=2.8,<3
pydantic-settings>=2.4
passlib[bcrypt]>=1.7
python-multipart>=0.0.9
httpx>=0.27
pytest>=8.2
ruff>=0.5
black>=24.0
isort>=5.13
mypy>=1.10
PyJWT>=2.9
